Get the ansible scripts deploying a functioning collectd & facette setup across two servers.  Still need to configure the mysql plugin and the deploy of it to the db servers
Also would be useful to refactor the scripts so collectd, facette,
collectd-server & collectd-client are roles which are applied to the
different servers.

This also includes an epel role that is dual licensed MIT and GPL2 which
installs the epel yum repository.  Yum packages are only used on the
health server, the rpms are included for the servers without direct network
connections